moved include-parameter-in-message.mjs to root and exported for use in sub-classes; updated docs and referenced
