+++ src/argument-invalid-error.mjs
@@ -1,6 +1,6 @@
 /* globals ArgumentMissingError ArgumentOutOfRangeError ArgumentTypeError mapErrorToHttpStatus */
 import { CommonError } from './common-error'
-import { includeParameterInMessage } from './lib/include-parameter-in-message'
+import { includeParameterInMessage } from './include-parameter-in-message'
 import { registerParent } from './map-error-to-http-status'
 import { translateValue } from './lib/translate-value'
