+++ src/argument-out-of-range-error.mjs
@@ -1,6 +1,6 @@
 /* globals ArgumentMissingError ArgumentTypeError CommonError */ // used in docs
 import { ArgumentInvalidError } from './argument-invalid-error'
-import { includeParameterInMessage } from './lib/include-parameter-in-message'
+import { includeParameterInMessage } from './include-parameter-in-message'
 import { registerParent } from './map-error-to-http-status'
 import { translateValue } from './lib/translate-value'
