+++ src/bad-credentials-error.mjs
@@ -1,6 +1,6 @@
 /* globals CommonError */ // ref in docs
 import { AuthError } from './auth-error'
-import { includeParameterInMessage } from './lib/include-parameter-in-message'
+import { includeParameterInMessage } from './include-parameter-in-message'
 import { registerParent } from './map-error-to-http-status'
 
 const myName = 'BadCredentialsError'
