+++ src/database-error.mjs
@@ -1,6 +1,6 @@
 /* globals RollbackError TransactionError */ // used in docs
 import { CommonError } from './common-error'
-import { includeParameterInMessage } from './lib/include-parameter-in-message'
+import { includeParameterInMessage } from './include-parameter-in-message'
 import { registerParent } from './map-error-to-http-status'
 
 const myName = 'DatabaseError'
