+++ src/index.mjs
@@ -36,6 +36,7 @@ export * from './unavailable-error'
 export * from './unique-constraint-violation-error'
 // Status management functions
 export * from './common-error-settings'
+export * from './include-parameter-in-message'
 export * from './mask-no-access-errors'
 export * from './map-error-to-http-status'
 export * from './map-http-status-to-name'
