+++ src/not-implemented-error.mjs
@@ -1,6 +1,6 @@
 /* global NotSupportedError UnavailableError */ // used in docs
 import { CommonError } from './common-error'
-import { includeParameterInMessage } from './lib/include-parameter-in-message'
+import { includeParameterInMessage } from './include-parameter-in-message'
 import { registerParent } from './map-error-to-http-status'
 
 const myName = 'NotImplementedError'
